+++ src/driver/drv_uart.h
@@ -1,18 +1,21 @@
 #pragma once
 
+#include <stdint.h>
+#include <stdbool.h>
+
 //---------------------------------------------------
 // Routines using UART port depending on config 
 // flag OBK_FLAG_USE_SECONDARY_UART
 // Note: using same buffer as routines below
 //--------------------------------------------------
-void UART_InitReceiveRingBuffer(int size);
+void UART_InitReceiveRingBuffer(uint32_t size);
 int UART_GetReceiveRingBufferSize();
 int UART_GetDataSize();
-byte UART_GetByte(int idx);
-void UART_ConsumeBytes(int idx);
-void UART_AppendByteToReceiveRingBuffer(int rc);
-void UART_SendByte(byte b);
-int UART_InitUART(int baud, int parity, bool hwflowc);
+uint8_t UART_GetByte(uint32_t idx);
+void UART_ConsumeBytes(uint32_t idx);
+void UART_AppendByteToReceiveRingBuffer(uint8_t rc);
+void UART_SendByte(uint8_t b);
+int UART_InitUART(uint32_t baud, uint32_t parity, bool hwflowc);
 void UART_AddCommands();
 void UART_RunEverySecond();
 
@@ -33,13 +36,13 @@ int UART_GetSelectedPortIndex();
 #define UART_PORT_INDEX_1 1
 //
 int UART_GetBufIndexFromPort(int aport);
-void UART_InitReceiveRingBufferEx(int auartindex, int size);
-void UART_AppendByteToReceiveRingBufferEx(int auartindex, int rc);
-int UART_GetReceiveRingBufferSizeEx(int auartindex);
-int UART_GetDataSizeEx(int auartindex);
-byte UART_GetByteEx(int auartindex, int idx);
-void UART_ConsumeBytesEx(int auartindex, int idx);
-void UART_SendByteEx(int auartindex, byte b);
-int UART_InitUARTEx(int auartindex, int baud, int parity, bool hwflowc);
-void UART_LogBufState(int auartindex);
+void UART_InitReceiveRingBufferEx(uint32_t auartindex, uint32_t size);
+void UART_AppendByteToReceiveRingBufferEx(uint32_t auartindex, uint8_t rc);
+int UART_GetReceiveRingBufferSizeEx(uint32_t auartindex);
+int UART_GetDataSizeEx(uint32_t auartindex);
+uint8_t UART_GetByteEx(uint32_t auartindex, uint32_t idx);
+void UART_ConsumeBytesEx(uint32_t auartindex, uint32_t idx);
+void UART_SendByteEx(uint32_t auartindex, uint8_t b);
+int UART_InitUARTEx(uint32_t auartindex, uint32_t baud, uint32_t parity, bool hwflowc);
+void UART_LogBufState(uint32_t auartindex);
